requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.1.0
beanie==1.23.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dateparser==1.2.0
dnspython==2.4.2
frozenlist==1.4.0
idna==3.6
lazy-model==0.2.0
magic-filter==1.0.12
motor==3.3.2
multidict==6.0.4
pycryptodome==3.19.0
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pymongo==4.6.1
python-binance==1.0.19
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
regex==2023.10.3
requests==2.31.0
six==1.16.0
toml==0.10.2
typing_extensions==4.8.0
tzlocal==5.2
ujson==5.8.0
urllib3==2.1.0
websockets==12.0
yarl==1.9.3
